move github event to listeners module
