+++ backend/event-listeners/src/bin/github-indexer/indexer/composite.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use derive_new::new;
-use event_listeners::domain::GithubEvent;
+use event_listeners::listeners::github::Event as GithubEvent;
 use futures::future::try_join_all;
 
 use super::{Indexable, Result};
