+++ backend/event-listeners/src/bin/github-indexer/indexer/guarded.rs
@@ -1,7 +1,7 @@
 use std::{future::Future, marker::PhantomData};
 
 use async_trait::async_trait;
-use event_listeners::domain::GithubEvent;
+use event_listeners::listeners::github::Event as GithubEvent;
 
 use super::{Indexable, Result};
