+++ backend/event-listeners/src/bin/github-indexer/indexer/mod.rs
@@ -18,7 +18,7 @@ use std::{
 
 use async_trait::async_trait;
 use error::Result;
-use event_listeners::domain::GithubEvent;
+use event_listeners::listeners::github::Event as GithubEvent;
 pub use repository::Repository;
 
 pub trait Stateful<Id> {
