+++ backend/event-listeners/src/bin/github-indexer/indexer/published.rs
@@ -2,7 +2,7 @@ use std::{marker::PhantomData, sync::Arc, time::Duration};
 
 use async_trait::async_trait;
 use domain::{Destination, Publisher};
-use event_listeners::{domain::GithubEvent, GITHUB_EVENTS_EXCHANGE};
+use event_listeners::{listeners::github::Event as GithubEvent, GITHUB_EVENTS_EXCHANGE};
 use infrastructure::amqp::UniqueMessage;
 
 use super::{Indexable, Result, Stateful};
