+++ backend/event-listeners/src/bin/github-indexer/indexer/user.rs
@@ -4,7 +4,7 @@ use async_trait::async_trait;
 use chrono::{DateTime, Utc};
 use derive_new::new;
 use domain::{GithubFetchService, GithubFullUser, GithubUserId};
-use event_listeners::{domain::GithubEvent, models::GithubUserIndexRepository};
+use event_listeners::{listeners::github::Event as GithubEvent, models::GithubUserIndexRepository};
 use serde::{Deserialize, Serialize};
 
 use super::{error::IgnoreErrors, hash, Result};
