+++ backend/event-listeners/src/bin/github-indexer/main.rs
@@ -3,7 +3,7 @@ use std::{sync::Arc, time::Duration};
 use anyhow::Result;
 use domain::{GithubRepoId, GithubUserId, LogErr};
 use dotenv::dotenv;
-use event_listeners::{domain::GithubEvent, Config};
+use event_listeners::{listeners::github::Event as GithubEvent, Config};
 use indexer::{
 	composite::Arced, guarded::Guarded, logged::Logged, published::Published,
 	with_state::WithState, Indexable, Indexer,
