--- backend/event-listeners/src/domain/github/mod.rs
@@ -1,2 +1,0 @@
-mod events;
-pub use events::Event;
