--- backend/event-listeners/src/domain/mod.rs
@@ -1,2 +1,0 @@
-mod github;
-pub use github::Event as GithubEvent;
