+++ backend/event-listeners/src/lib.rs
@@ -4,7 +4,6 @@ extern crate diesel;
 #[macro_use]
 extern crate derive;
 
-pub mod domain;
 pub mod listeners;
 pub mod models;
